examples/python-examples/profiling/pyproject.toml
[project]
name = "profiling"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyroscope-io==0.8.8",
    "Flask>=3.1.1",
    "requests>=2.32.5",
]

[tool.black]
line-length = 120

[tool.isort]
profile = "black" # 兼容 black
line_length = 120
